backend/requirements.txt
anthropic>=0.39.0
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
lxml>=5.3.0
python-dotenv>=1.0.1
